ethereum_api_python/requirements.txt
web3==6.9.0
python-dotenv==1.0.0